requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
httpx
requests
beautifulsoup4
lxml
playwright
playwright-stealth
railtracks
railtracks[cli]
openai
jinja2
audioop-lts; python_version >= "3.13"
numpy
twilio>=9.0.0
boto3>=1.34.0
mangum
